test_backend/requirements.txt
aiohttp==3.8.1
aiohttp-sse-client==0.2.1
aiosignal==1.2.0
alabaster==0.7.12
antiorm==1.2.1
asgiref==3.5.2
astroid==2.11.6
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
blank==0.1.0
cached-property==1.5.2
Cerberus==1.3.4
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
codecov==2.1.12
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==3.4.8
db==0.1.1
db-sqlite3==0.0.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
Django==3.2.13
django-appconf==1.0.5
django-compressor==4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-model-utils==4.2.0
django-polaris==2.3.1
django-rest-swagger==2.2.0
django-sass-processor==1.2.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.17.1
drf-yasg==1.20.0
filelock==3.7.1
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
imagesize==1.4.1
importlib-metadata==4.12.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Markdown==3.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
mnemonic==0.20
multidict==6.0.2
oauthlib==3.2.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==20.9
pbr==5.9.0
pep517==0.12.0
Pillow==7.2.0
pip-shims==0.7.2
platformdirs==2.5.2
plette==0.2.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.14.4
pylint-plugin-utils==0.7
PyNaCl==1.5.0
pyparsing==2.4.7
pytest==7.1.2
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
requirementslib==1.6.6
rjsmin==1.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
Sphinx==1.8.6
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
sqlparse==0.4.2
stellar-base-sseclient==0.0.21
stellar-sdk==8.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
typeguard==2.13.3
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.1
virtualenv-clone==0.5.7
vistir==0.5.2
whitenoise==5.3.0
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
